ALPHA 1.1: AUDIO IMPLEMENTATION
Implemented audio:
Implemented AudioMultiTracks and AudioMultiTrackCollections.
AudioMultiTracks are classes that contain a pointer to a std::vector containing audio sample instances, that can be used to play a random sample instance.
AudioMultriTrackCollections are used to conveniently store AudioMultiTracks, and are mainly used by Tiles, GroundTiles, and other objects that need to produce different sounds.