ALPHA 1.3: QUESTS
Implemented the Quest System
Quests are imported as JSON data.
Quests currently have no completion rewards.
For a Quest to be unlocked, all its quest requirements must be completed.
For a Quest to be completed, after it has been unlocked, all of its other requirements must be completed.
There are 4 types of quest requirements:
1. Quest completion requirements
2. Tile mine requirements
3. Ground tile mine requirements
4. Entity kill requirements