ALPHA 1.4: CRAFTING
Implemented Crafting GUI for the Player
The player can now open the crafting GUI by pressing 'C', scroll up or down and click on a recipe, and select how many times to craft.
Changed GUI's and UIComponent's methods to now return bool (events).
The methods should return true when the event was consumed, and false if it wasn't and should instead be propagated to the parent (if the click event should be forwarded to the component/gui(s) "behind")
Also created a configuration file (Config) that introduces the gameconfig namespace that contains variables with different colors that can be used in GUIs.
Created TextUtils file and namespace that contains utility functions for drawing text (currently unused/WIP; may be removed in the future)