ALPHA 1.12: Menus
When the game is opened, the new MainMenuGUI is shown, and the menu theme is played on an infinite loop.
Then, the player can press "Play" and either create or load a world, view the credits menu, or press "Quit" to exit the game.
Also, when the user has loaded or created a new world, he can also press escape to open the new pause menu, to save or exit the world.
Pressing CTRL+S no longer saves the world.